+++ packages/dataviews/src/dataviews-layouts/grid/preview-size-picker.tsx
@@ -13,6 +13,14 @@ import type { ViewGrid } from '../../types';
 
 const imageSizes = [
 	{
+		value: 120,
+		breakpoint: 1,
+	},
+	{
+		value: 170,
+		breakpoint: 1,
+	},
+	{
 		value: 230,
 		breakpoint: 1,
 	},
@@ -34,24 +42,17 @@ export default function PreviewSizePicker() {
 	const context = useContext( DataViewsContext );
 	const view = context.view as ViewGrid;
 
-	if ( context.containerWidth < 588 ) {
-		return null;
-	}
-
 	const breakValues = imageSizes.filter( ( size ) => {
 		return context.containerWidth >= size.breakpoint;
 	} );
 
+	const layoutPreviewSize = view.layout?.previewSize ?? 230; // Default to the third smallest size if no preview size is set.
 	// If the container has resized and the set preview size is no longer available,
 	// we reset it to the next smallest size.
-	const previewSizeToUse = view.layout?.previewSize
-		? breakValues
-				.map( ( size, index ) => ( { ...size, index } ) )
-				.filter(
-					( size ) => size.value <= ( view.layout?.previewSize ?? 0 ) // We know the view.layout?.previewSize exists at this point but the linter doesn't seem to.
-				)
-				.sort( ( a, b ) => b.value - a.value )[ 0 ].index
-		: 0;
+	const previewSizeToUse = breakValues
+		.map( ( size, index ) => ( { ...size, index } ) )
+		.filter( ( size ) => size.value <= layoutPreviewSize )
+		.sort( ( a, b ) => b.value - a.value )[ 0 ].index;
 
 	const marks = breakValues.map( ( size, index ) => {
 		return {