+++ packages/dataviews/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Add `enableMoving` option to the `table` layout to allow or disallow column moving left and right. [#71120](https://github.com/WordPress/gutenberg/pull/71120)
 
+### Enhancements
+
+- Add two smaller sizs to the grid layout ([#71077](https://github.com/WordPress/gutenberg/pull/71077)).
+
 ## 6.0.0 (2025-08-07)
 
 ### Breaking changes